+++ apps/backend/src/services/products/remove.service.ts
@@ -1,7 +1,23 @@
 import { prisma } from "@repo/database";
 import { toProductWithInventoryItemsDto } from "../utils/product.mapper.js";
+import { AppError } from "../../errors/AppError.js";
 
 export const remove = async (id: string) => {
+  const product = await prisma.product.findUnique({
+    where: {
+      id,
+    },
+    select: {
+      id: true,
+    },
+  });
+
+  if (!product)
+    throw new AppError({
+      message: "Product not found",
+      code: "NOT_FOUND",
+    });
+
   const result = await prisma.product.delete({
     where: {
       id,
